+++ PreferencePaneController.m
@@ -424,11 +424,11 @@ extern BrowserController	*browserWindow;
 }
 
 //TableViews Data source
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView{
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView{
 	return 0;
 }
 
-- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex{
+- (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(NSInteger)rowIndex{
 	return nil;
 }
 @end
